requirements-dev.txt
-r requirements.txt
Sphinx==1.6.4
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
numpydoc==0.7.0
nose==1.3.7
